+++ NEWS.md
@@ -1,5 +1,12 @@
 # *News*
 
+# zipcodeR 0.3.0
+- Added `search_radius()` function to allow searching for ZIP codes around a radius of lat / lon coordinates.
+- Added `zip_distance()` function for calculating the distance between ZIP codes using their centroids.
+- Added `geocode_zip()` function that returns the lat / lng centroid of a given ZIP code.
+- Added `normalize_zip()` function for normalizing messy ZIP code input (Contributed by Claus Wilke).
+- The `reverse_zipcode()` function has been updated to return a blank row for invalid ZIP codes with no matches in the zip code database.
+
 # zipcodeR 0.2.0
 - `search_county()` function now allows for approximate matching of county names using agrep (Andre Mikulec)
 - `search_state()` is now vectorized and will accept a vector of state abbreviations